fix: use new blanket events & clean up eventing
Before, there was one global eventHandler for key presses, but if multiple inline dialogs got opened at the same time, the key handler was removed for all of them.
Also, the blanket activation event is only registered if there is a blanket and not registered on the window any more, but on the blanket component itself.

affects: ak-blanket