refactor: Updated style() API to take vdom as its first argument.
affects: ak-theme, akutil-common

The problem is that akutil-common is importing Skate to use the vdom module. Other components import

Skate and use their own vdom modules. Even if it's the same version of Skate, separate instances of

Incremental DOM don't play well because they store information locally. Because of this, we have to

pass in the instance of the virtual DOM module that we want the helper to use.

AK-39