refactor(component): ak-page no longer uses explicit width
affects: ak-page

ak-page now depends on the navigation taking up width - this was achieved in ak-navigation through

the user of a "spacer element" that takes up the same width as the `position: fixed` navigation

element

BREAKING CHANGE: the `navigation` is no longer position fixed, and `page.navigationWidth` was

removed. This is because it is no longer needed for the implementation of the navigation width.