fix(exclusiveRight): right part of a link is not a part of the link
affects: ak-editor-schema

When I select some text, hit the hyperlink button, enter link into the dialog box and hit enter.

Hyperlink should be added. If you continue typing at the end of hyperlink(space or not) or in the

front, it should not be part of hyperlink.

If you continue typing at the middle of hyperlink, it should be part of the hyperlink