fix(component): refactor container hidden bug to avoid the search drawer open / close bug
affects: ak-navigation

as far as I can tell, how it was before should have worked fine. However, there was something funny

with toggling the overflow: hidden / visible of the navigation when the container was hidden. This

approach just doesn't have overflow: hidden when the container is hidden, because it's not really

neede.