refactor(component): Refactored ComponentBase to now require Component and prop to be injected in to
affects: ak-component-base

There is an issue with the renderer used in ComponentBase and whatever tries to extend it not

matching, so we inject them into the default export to create the class on the fly instead.

BREAKING CHANGE: The default export from ak-component-base is now a function used to inject the

Component and prop dependencies to create the ComponentBase class. See the readme.md for more

information.