chore: Pin tslint and tslint-react.
Updating tslint and tslint-react broke the build because it generated a lot of linting errors. I've
put a patch release fuzzy version on here so we don't break on the next minors, but at least get
patch updates if released. We should figure out what's wrong and update soon, though.

References AK-1423