fix(component): update mocha, refactor uses of this.skip() in tests.
affects: @atlaskit/editor-bitbucket

It seems that we were relying on non-existent API for skipping tests while supplying a reason

description with this.skip('reason'). The @types/mocha supported that but it seems that Mocha itself

doesn't, so we need to remove the argument from the call to prevent TypeScript from complaining.