refactor(component): Change the behaviour of the onResize handler's arguments
affects: @atlaskit/navigation

This is a relatively subtle change in the behaviour of the argument to onResize. If you are passing

the arguments from onResize directly to Navigation.width and Navigation.isOpen, you should notice no

change.

Old paramaters to onResize, by dragged width:
                            Dragged width
|  args  | < 64          | 64-174 | 174-284 | 284-334 | > 334 |
|--------|---------------|--------|---------|---------|-------|
| isOpen | F             | F      | T       | T       | T     |
| width  | undef         | undef  | undef   | undef   | > 334 |
|        |               |        |         |         |       |

New parameters to onResize, by dragged width:
                            Dragged Width
|  args  | < 64          | 64-174 | 174-284 | 284-334 | > 334 |
|--------|---------------|--------|---------|---------|-------|
| isOpen | F             | F      | T       | T       | T     |
| width  | 64            | 64     | 284     | 284-334 | > 334 |
|        |               |        |         |         |       |

This is coupled with two differences in rendered width calculation:

1) Before, if you passed in <Navigation isOpen={true} width={174 < width < 284} />, `width` would
be the rendered width. Now, 284px will be the rendered width.

2) Before, if you passed in <Navigation isOpen={true} width={284 < width < 334 />, 284px would
be the rendered width. Now, `width` will be the rendered width.

This means that if you are passing `width` and `isOpen` directly from Navigation.onResize back to
Navigation as props, you should have the intended interaction behaviour.

BREAKING CHANGE: Changed the arguments given to onResize so that width is always available as a

parameter. Also remove the 50px window of snapping to close down to 284px.