refactor(component): change the width of Navigation from 284px to 304px
affects: @atlaskit/navigation

Navigation width is now 64px+240px = 304px total.

BREAKING CHANGE:

Increased the total width of Navigation in the isOpen state from 284px to 304px. Increased the width

of ContainerNavigation from 220px to 304px. It is possible that this will not require any code

changes, it will change the default open state width though, and the widths to which the Navigation

will snap open