test(refactor): used members() instead of deepEqual() for asserting icon changes.
affects: @atlaskit/icon

I'm unsure if this was an issue before, or is exposed by updating the build pipeline, but it's
interesting why this wasn't a problem before. It seems like it should have been. Nonetheless, I've
ensured that changing the members of the compare array will, in fact, fail the test, so this works
moving forward.