feat(component): no longer rendering the Global navigation or drawers when they are not visible
affects: @atlaskit/navigation

BREAKING CHANGE:

Previously the navigation always rendered the global navigation and drawers even when they where not

visible. This had focus and performance implications. It was also previously hard to know when your

drawer was visible. Now it is simple - things are mounted when they are visible and not mounted when

they are not.

ISSUES CLOSED: AK-2151