feat(component): disable spell checking for SearchInput text field
affects: @atlaskit/navigation

When spell checking is not explicitly disabled, some browsers such as Chrome enable it and show red
squiggly lines for words. For something like a search text field, this is unwanted as it will
squiggle search terms such as names or technical terms that are not in the dictionary.

ISSUES CLOSED: AK-2240