feat(component): stop rendering elements that are not visible. Simplify and centralise the logic con
affects: @atlaskit/navigation

Previously the Navigation components always rendered everything in the Navigation tree all the time.

This has been changed so that only components that are visible to the user will be rendered by

React.

Issues this resolves:

- focus management

- more sane api around components: they can now

just know that if they are mounted that they are visible

- performance: react no longer needs to

worry about things that are not visible

- positioning: there was a lot of craziness in

implementation and api to maintain the idea that elements where always rendered but sometimes off

screen or not visible. This change makes the positioning logic easier to follow.

This change also

reduces the amount of props required to control the width of the navigation. Because things that are

not visible are not rendered there is not as much fine grained width control required.

Prop

changes

ContainerNavigation

areGlobalActionsVisible: (rename) =>

showGlobalPrimaryActions

isCollapsed: (new) whether or not the container is

collapsed

shouldAnimate: (removed) - now managed by Navigation

width: (removed) - now managed by

Navigation

offsetX: (removed) - now managed by Navigation

BREAKING CHANGE:

Nested components are now only mounted when visible. Prop changes to ContainerNavigation.

ISSUES CLOSED: AK-2151