fix(component): fix a bug, clean up unused class and remove an unnecessary element.
affects: @atlaskit/media-card

fixed a bug where the animation would start out-of-align when the list was wider than the width of a

card.

cleaned up an unused class for the previously removed load-more-button.

removed the span

added by CSSTransitionGroup and applied it to the CardListWrapper to reduce the depth of elements