refactor(refactor): refactor dynamic-table as part of styled-components refactor
affects: @atlaskit/dynamic-table

While the component was previously converted to styled-compoents, this aims to bring the component
in line with other refactors, matching the shape we have settled on. It also changes the name of the
exports for consistency and clarity.

BREAKING CHANGE:
DynamicTable export renamed to DynamicTableStateless, for consistency with other packages.

ISSUES CLOSED: #AK-2385