fix(component): cleaning up how Navigation does positioning
affects: @atlaskit/navigation

Previously Navigation was using a lot of techniques that made layering, positioning and zindexing
really hard.
This change strips back and untangles the positioning and layering strategy of
Navigation to be much simplier and more robust.
Even better: this was done with no api changes

ISSUES CLOSED: AK-2420