feat(component): all colours in navigation are now based on themes
affects: @atlaskit/navigation

This is the first piece of work required to allow custom theming of navigation.
This change
takes
the existing appearance props on GlobalNavigation and ContainerNavigation and maps them to
preset
theme definitions.
A future pull request will allow consumers to generate and provide their
own
themes.
This change involves the removal of appearance props on some children of
ContainerNavigation
and GlobalNavigation as they are now controlled by the appearance prop on
ContainerNavigation and
GlobalNavigation respectively.
This change involves the removal of
appearance props on some
children of ContainerNavigation and GlobalNavigation as they are now
controlled by the appearance
prop on ContainerNavigation and GlobalNavigation respectively.
This
change also involves the removal
of the isCollapsed prop from NavigationItemGroup. Whether a
NavigationItem is collapsed is now
controlled only by whether it has its own isCollapsed prop
set.
Previously you could set the
isCollapsed on either the group or the item which was confusing.

BREAKING CHANGE:
This change is a breaking change because it involves: 1. making the appearance prop on some children
of ContainerNavigation and GlobalNavigation unused which can have unexpected consequences if you
where doing strange things with appearance props. 2. A context dependency is introduced in
components that previously did not have context dependencies. If you are using Navigation components
within Navigation then everything should be fine. However, if you are using them outside of
Navigation there will now be a dependency on theme in the context. Tooling for this use case has not
been created because it is not the designed us case.

ISSUES CLOSED: AK-2433 AK-2483