feat(component): editor-bitbucket: Add support for new BB emoji markup
affects: @atlaskit/editor-bitbucket

editor-bitbucket currently parsing Bitbucket-rendered HTML markup to
generate the initial state for the editor. Part of this parsing process
involves inspecting img.emoji elements to find data about emoji that the
editor should render. Bitbucket is now changing the format of it's emoji
markup and the editor-bitbucket component must be modified to be
compatible.

The current method involves inspecting the src attribute of the <img>
and parsing the emoji short name out of the URL. The new <img> format
does not include the emoji short name in the file path but does contain
a new 'data-emoji-short-name' attribute which contains the short name
for the emoji.

This change jmodifies the parsing behavior to check for the new data
attribute and, if found, use it. Otherwise, the component will fall back
on the original behavior of parsing the <img> src attribute. This should
allow the editor to support both formats until Bitbucket eliminates the
old format completely at a later date.