feat(component): add ascii->emoji map to EmojiResource and EmojiRepository
affects: @atlaskit/emoji

Store mapping of emoji ascii representation to EmojiDescription to be able to
look up emojis based on their ascii representations. This will be used to bring
an ascii match to the top of the typeahead when sorting.

BREAKING CHANGE:
Added required getAsciiMap() method to EmojiProvider. Consumers will need to
implement it in their concrete classes.