feat(component): keep the leading colon in the query being passed to the typeahead
affects: @atlaskit/editor-core

Keeping the leading colon in the query so that it can be used by the typeahead to match against
ascii representations of emojis that start with a colon. It doesn't have any effect on the beahviour
of js-search so same filtering as before will occur.

ISSUES CLOSED: FS-916