feat(component): add ascii emoji input rule to automatically match and convert ascii representations
affects: @atlaskit/editor-core, @atlaskit/util-data-test

The new input rule matches short words against possible ascii representations and automatically
converts them. There are 2 uses cases dependeing on if the ascii represenation starts with a colon
or not. In the former case, the ascii is only converted if when followed by a space. In the latter
scenario, the ascii is converted as soon as a match is found. This ensures that it doesn't interfere
with the typeahead behaviour.

ISSUES CLOSED: FS-916