fix(component): fix error handling on website build
affects: @atlaskit/website

When a docs/index.js file is not found in a component, a warning will be printed, but the build will
continue, the component will be included in the website with a link out to the storybook.
When an
error is thrown that does not relate to docs/index.js not existing, the error will be thrown and the
build stopped.

ISSUES CLOSED: #2624