feat(component): introduce the new method findById(String) to EmojiProvider
affects: @atlaskit/emoji

When converting native emoji to our Emoji representation the only identifier we have for the emoji
is the unicode code point, which is also the unique id.
This change makes it possible to look up a
Emoji node based on only this information. Previously an optional shortname and an id were required.

BREAKING CHANGE:
The introduction of findById(String) to EmojiProvider is a breaking change.

ISSUES CLOSED: FS-935