feat(component): global secondary items are now visible when the navigation is collapsed
affects: @atlaskit/navigation

BREAKING CHANGE:
The ContainerNavigation.showGlobalPrimaryActions prop has been renamed to
ContainerNavigation.showGlobalActions. This is only a breaking change if you are explicitly
rendering ContainerNavigation via the named export. If you are using the default Navigation export
only, this is not a breaking change for you.

ISSUES CLOSED: AK-2591