fix(component): emoji typeahead now shows up after a hardBreak node
affects: @atlaskit/editor-core

A previous change to the emoji input rule's regex prevented the ability for it to trigger after a
hard break. PM uses a special Unicode character to represent a leaf node which was not being
matched. The new regex changes the behaviour such that the emoji input rule is now triggered if the
colon character is preceded by a leaf node. This currenlty includes hardBreak, emoji, mention and
rule (although rule is always in its own paragraph node).

ISSUES CLOSED: FS-1113