feat(component): now auto-convert ascii emoji starting with a colon to the matching emoji when follo
affects: @atlaskit/editor-core

Make the assumption that if users type a space after an emoji query matching an emoticon they meant
to have it converted to its matching emoji. Considering that emoticons are usually short and often
made of punctuation characters, this assumption should be correct almost all the time. For the rare
occasion where it's users still have the option to undo.

ISSUES CLOSED: FS-1120