feat(component): using css transitions rather than physics for minor movements to greatly increase p
affects: @atlaskit/drag-and-drop

css transitions are now used to move items out of the way of a dragging item.
No longer applying
physics to the item that is being dragged with a mouse - it should move instantly.
Cleanup of types
for provided.draggableStyle.

BREAKING CHANGE:
Minor changes to Draggable > Provided > DraggableStyle. In most cases upgrading will involve no
changes