fix(component): don't return non-searchable emoji in searches (and 'getAll') calls
affects: @atlaskit/emoji

Some Emoji can be modified with a gender and/or a skin tone. As an example, a police officer
can be
male or female. We don't want the non-gendered police officer to
appear in the typeahead or picker.
We want a user to only
be presented with the male and female options. The, third non-gendered
option tends to duplicate the male option in most cases.

BREAKING CHANGE:
The EmojiDescription and EmojiServiceDescription interfaces have an additional mandatory property.

ISSUES CLOSED: FS-1171