fix(component): no user selection in mention list should result in the top item always being selecte
affects: @atlaskit/mention

Previously if a subsequent list of mentions contained a mention that was the first item in the
previous list
then that item would appear selected, even if it was no longer the top of the list.
This behaviour is now only
applied if a user has deliberately selected a particular item in a list.

ISSUES CLOSED: FS-1178