feat(component): generic item component to be composed into other components
affects: @atlaskit/item

This component is intended to be used declaratively, inside dropdown-menu/droplist/navigation
initially. It is primarily concerned with layout, and higher-order components can be created to
apply additional functionality to this generic Item component. There is also a basic ItemGroup named
export component, which again can be used in conjunction with HOCs.

ISSUES CLOSED: AK-2642