fix(component): nav headers refactored and new stories added
affects: @atlaskit/navigation

The nav header component did not support inline dialogs and dropdown menus very well. It has better
support, but the containerHeaderComponent now needs to take a function that returns an array of
components, as they are now flex children.

BREAKING CHANGE:
containerHeaderComponent now needs to return an array of components