feat(component): refactor of Item and Navigation to support Project Switcher dropdown menus
affects: @atlaskit/item, @atlaskit/navigation

BREAKING CHANGE:
ContainerTitleDropdown rebuilt to use Item and Dropdown under the hood. No longer accepts a
LinkComponent prop because it is always a button for a dropdown menu. It also accepts children as
the dropdown menu items. Essentially it calls out to DropdownMenu with props that are specific to
Navigation Container Headers.