feat(component): emoticons now auto-convert after an opening rounded bracket
affects: @atlaskit/editor-core

Entering an emoticon like ;) after an opening rounded bracket now converts it to an emoji. The
implementation is not ideal as it makes a special case for (y), the only emoticon starting with an
opening rounded bracket. The regex could probably be improved to cater for any cases but for now we
need to easily differentiate between (y) and (;)

ISSUES CLOSED: FS-1247