refactor(component): fix white border around avatar and adjust alignment of MentionItem's contents
affects: @atlaskit/mention

The new version of Avatar adds a 4x (for medium size) border around the avatar
which is white by default. It was visible when a mention row was highlighted
with its grey background. The colour of the border is now set to same colour as
the background of the row.

The addition of the border also caused misalignment of elements in the mention
item row.

ISSUES CLOSED: FS-1309