fix(component): nav drawer content fills drawer height, to allow for vertically-centered layouts
affects: @atlaskit/navigation

BREAKING CHANGE:
DrawerContent is now a flex column, so anything you put inside a drawer will behave like a flex
child. In most cases this won't have any discernable impact and this will be an easy update. One
instance where this might catch you out is if you have an inline element as a direct child of a
Drawer - it will be stretched to fill the width of the container. To work around this simply wrap
everything inside the Drawer in a div.

ISSUES CLOSED: AK-3443, DONUT-789