fix(component): make mention lozenge 20px high to match rendering line height
affects: @atlaskit/mention

Because of the padding and border width, the mention lozenge could be up to 21px high. This caused
an issue when rendering multiple mention vertically side by side as they would overlap each other.
The design spec actually defines the height of the mention lozenge as 20px.

ISSUES CLOSED: FS-1160