fix(component): fix error where mark type groups were not being properly filtered out
affects: @atlaskit/editor-core

We were incorrectly filtering the \`excludes\` attribute on mark-type definitions to exclude any
groups that did not exist as a mark-type name, rather than by the groups declared by the rest of the
mark-types. This fixes a number of issues with being able to apply styles to inline code-blocks &
other excludes violations.

ISSUES CLOSED: ED-2805