feat(component): add support for detecting mark-types in a selection
affects: @atlaskit/editor-core

Our existing approach to detecting whether we could insert a mark at the current selection only
considered the $from position of the selection, rather than the whole selection. As a result, you
could make a selection from a valid node to a node with an excluded mark-type and it would still say
this was valid. We also now check whether a mark is excluded bi-directionally (as a mark can exclude
another, and vice-versa). Previously we would only check if the given mark excluded the existing
marks (which didn't tell us if the existing marks excluded the given mark)