fix(component): check if link is compatible with existing marks before adding link
affects: @atlaskit/editor-core

Previously we would check if adding a link to this node was valid, however as inline-code is a mark
we would be checking if adding a link to a text node was valid (which it is). We now do a search to
see whether the entire selection is compatible with the link marktype.

ISSUES CLOSED: ED-2218