fix(component): change global nav items to use styled buttons or styled anchor
affects: @atlaskit/navigation

Global nav now either renders a button with styles, or an anchor tag with styles, instead of
sometimes wrapping a styled span in an anchor, making it more accessible. This has implications for
dom-rendered of custom link components, however for our base use-cases (react-router), storybooks
render the same before and after. This will need to be monitored.