fix(component): fix issue where layer would flicker when Popper updated positioning
affects: @atlaskit/layer

For example, when a dropdown-menu's positioning needed to be "flipped" on initial mount, the layer
would render in the standard position and then re-render in the flipped position, causing a flicker
effect. Layer now renders the content with opacity:0 until Popper.js has had a chance to detect the
correct position.

ISSUES CLOSED: AK-3600