refactor(component): set the imageUploadHandler using providers
affects: @atlaskit/editor-core

Refactor the imageUploadPlugin to take advantage of the provider architecture in place to set it's
event handler. This is to simplify the integration with Bitbucket (as we'd otherwise have to have an
exception for this plugin, to set the handler)

BREAKING CHANGE:
The imageUploadHandler prop has been removed in favor of imageUploadProvider.