refactor(stories): fix the "no access" mention lozenge story that had its tooltip hidden
affects: @atlaskit/mention

The problem is that the tooltip is wrapped on 2 lines and thus the top of the tooltip would be
clipped by the viewport. Because of that the tooltip component detects it and flips the position
from the right to left. However that makes the matter worse as then the tooltip would be completely
hidden behind the viewport to the left of the lozenge.